\name{Trace}
\Rdversion{1.1}
\alias{Trace}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Trace of a square matrix
}
\description{
Compute the trace of a square matrix.
}
\usage{
Trace(X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
a square matrix.
}
}

\value{
Returns a scalar that is the sum of the diagonal elements of \code{X}.
}
\references{
Schott, J. (1997) Matrix Analysis for Statistics. Wiley.
}
\author{
Kofi P. Adragni <kofi.adragni@gmail.com>
}
